+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/ui/components/zoomimage/SketchZoomAsyncImage.kt
@@ -39,6 +39,7 @@ import com.github.panpf.sketch.LocalPlatformContext
 import com.github.panpf.sketch.PainterState
 import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.internal.AsyncImageContent
+import com.github.panpf.sketch.name
 import com.github.panpf.sketch.rememberAsyncImagePainter
 import com.github.panpf.sketch.rememberAsyncImageState
 import com.github.panpf.sketch.request.ImageRequest
@@ -227,7 +228,9 @@ private fun onPainterState(
     request: ImageRequest,
     loadState: PainterState,
 ) {
-    zoomState.zoomable.logger.d { "SketchZoomAsyncImage. onPainterState. state=${loadState.name}. uri='${request.uri}'" }
+    zoomState.zoomable.logger.d {
+        "SketchZoomAsyncImage. onPainterState. state=${loadState.name}. uri='${request.uri}'"
+    }
     val painterSize = loadState.painter
         ?.intrinsicSize
         ?.takeIf { it.isSpecified }
@@ -246,14 +249,6 @@ private fun onPainterState(
         }
     }
 }
-
-private val PainterState.name: String
-    get() = when (this) {
-        is PainterState.Loading -> "Loading"
-        is PainterState.Success -> "Success"
-        is PainterState.Error -> "Error"
-        is PainterState.Empty -> "Empty"
-    }
 
 private fun Size.roundToIntSize(): IntSize {
     return IntSize(width.roundToInt(), height.roundToInt())