+++ sketch-compose-core/src/commonMain/kotlin/com/github/panpf/sketch/PainterState.kt
@@ -2,7 +2,6 @@ package com.github.panpf.sketch
 
 import androidx.compose.runtime.Stable
 import androidx.compose.ui.graphics.painter.Painter
-import com.github.panpf.sketch.PainterState.Empty
 import com.github.panpf.sketch.PainterState.Error
 import com.github.panpf.sketch.PainterState.Loading
 import com.github.panpf.sketch.PainterState.Success
@@ -18,11 +17,6 @@ sealed interface PainterState {
 
     /** The current painter being drawn by [AsyncImagePainter]. */
     val painter: Painter?
-
-    /** The request has not been started. */
-    data object Empty : PainterState {
-        override val painter: Painter? get() = null
-    }
 
     /** The request is in-progress. */
     data class Loading(
@@ -51,6 +45,5 @@ val PainterState.name: String
         is Loading -> "Loading"
         is Success -> "Success"
         is Error -> "Error"
-        is Empty -> "Empty"
-        else -> "Unknown PainterState: $this"
+        else -> this.toString()
     }